requirements.txt
streamlit==1.28.0
google-generativeai==0.3.0
requests==2.31.0
python-dotenv==1.0.0
python-dateutil==2.8.2
typing-extensions==4.8.0